profile/fmp/central/src/androidTest/java/org/im97mori/ble/profile/fmp/central/db/FindMeProfileBondedDatabaseHelperTest.java
package org.im97mori.ble.profile.fmp.central.db;

import androidx.test.core.app.ApplicationProvider;

import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;

public class FindMeProfileBondedDatabaseHelperTest {

    @Test
    public void test_constructor_00001() {
        assertNotNull(FindMeProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()));
    }

    @Test
    public void test_constructor_00002() {
        assertEquals(FindMeProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()), FindMeProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()));
    }

    @Test
    public void test_getProfileName_00001() {
        assertEquals(FindMeProfileBondedDatabaseHelper.PROFILE_NAME, FindMeProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()).getProfileName());
    }
}
